@@ -1,17 +1,20 @@
 import React, { useEffect } from "react";
 import { useDispatch } from "react-redux";
 
-import { HashRouter as BrowserRouter, Route, Switch } from "react-router-dom";
+// import {BrowserRouter, Route, Switch} from "react-router-dom";
+// electron needs HashRouter
+import { HashRouter as BrowserRouter, Route, Routes } from "react-router-dom";
 import { ErrorNotification } from "../components/molecules/notifications/ErrorNotification";
 import { ControlPanelController } from "../components/organisms/controllPanel/presentation/ControlPanelController";
 import { SidePanelController } from "../components/organisms/sidePanel/SidePanelController";
+import "../css/Davit.css";
 import { ActorModelController } from "../domains/actor/ActorModelController";
 import { DataModelController } from "../domains/datamodel/DataModelController";
 import { FlowChartController } from "../domains/overview/flowChartModel/FlowChartController";
 import { TableModelController } from "../domains/overview/tableModel/presentation/TableModelController";
 import { GlobalActions } from "../slices/GlobalSlice";
 import { MasterDataActions } from "../slices/MasterDataSlice";
-import "./Davit.css";
+import { useAppDispatch } from "../store";
 
 export const ModuleRoutes = {
     home: "/",
@@ -23,7 +26,7 @@ export const ModuleRoutes = {
 
 // inital data load from backend.
 export function Davit() {
-    const dispatch = useDispatch();
+    const dispatch = useAppDispatch();
     useEffect(() => {
         dispatch(MasterDataActions.loadAll());
         dispatch(GlobalActions.loadActorZoomFromBackend());
@@ -32,10 +35,11 @@ export function Davit() {
 
     return (
         <BrowserRouter>
-            <Switch>
-                <Route exact
+            <Routes>
+                <Route
                        path={ModuleRoutes.home}
-                >
+                       element={
+                    <>
                     <div className="davitGridContainer">
                         <ControlPanelController />
                         <ActorModelController />
@@ -45,36 +49,55 @@ export function Davit() {
                         <TableModelController />
                         <ErrorNotification />
                     </div>
+                    </>
+                       }
+                       >
                 </Route>
-                <Route exact
+                {/* <Route
                        path={ModuleRoutes.actor}
                 >
-                    <div className="componentPage">
-                        <ActorModelController />
+                    <>
+                    <div className="Carv2">
+                        <div className="componentPage">
+                            <ActorModelController fullScreen />
+                        </div>
                     </div>
+                    </>
                 </Route>
-                <Route exact
+                <Route
                        path={ModuleRoutes.data}
                 >
-                    <div className="componentPage">
-                        <DataModelController />
+                    <>
+                    <div className="Carv2">
+                        <div className="componentPage">
+                            <DataModelController fullScreen />
+                        </div>
                     </div>
+                    </>
                 </Route>
-                <Route exact
+                <Route
                        path={ModuleRoutes.table}
                 >
-                    <div className="componentPage">
-                        <TableModelController />
+                    <>
+                    <div className="Carv2">
+                        <div className="componentPage">
+                            <TableModelController fullScreen />
+                        </div>
                     </div>
+                    </>
                 </Route>
-                <Route exact
+                <Route
                        path={ModuleRoutes.flowChart}
                 >
-                    <div className="componentPage">
-                        <FlowChartController />
+                    <>
+                    <div className="Carv2">
+                        <div className="componentPage">
+                            <FlowChartController fullScreen />
+                        </div>
                     </div>
-                </Route>
-            </Switch>
+                    </>
+                </Route> */}
+            </Routes>
         </BrowserRouter>
     );
 }
